renames lambda folder to functions
refactors all from 'lambda' to 'functions'
