+++ .env.example
@@ -1,1 +1,1 @@
-GATSBY_LAMBDA_URL=
+GATSBY_FUNCTIONS_URL=
