+++ functions/src/createProof.ts
