+++ functions/src/getExample.ts
