+++ functions/src/getRewardData.ts
