+++ functions/src/getTotalAssetsTokenized.ts
