+++ functions/src/handleCreateProof.ts
