+++ functions/src/subgraphUtils.ts
